refactor: service 내 기능별 폴더 정리
